Crate rusoto_kinesis_video_media

If you're using the service, you're probably looking for KinesisVideoMediaClient and KinesisVideoMedia.

Structs

GetMediaInput

GetMediaOutput

KinesisVideoMediaClient
A client for the Kinesis Video Media API.

StartSelector

Identifies the chunk on the Kinesis video stream where you want the GetMedia API to start returning media data. You have the following options to identify the starting chunk:

Choose the latest (or oldest) chunk.

Identify a specific chunk. You can identify a specific chunk either by providing a fragment number or timestamp (server or producer).

Each chunk's metadata includes a continuation token as a Matroska (MKV) tag (AWSKINESISVIDEOCONTINUATION_TOKEN). If your previous GetMedia request terminated, you can use this tag value in your next GetMedia request. The API then starts returning chunks starting where the last API ended.

Enums

GetMediaError
Errors returned by GetMedia

Traits

KinesisVideoMedia
Trait representing the capabilities of the Kinesis Video Media API. Kinesis Video Media clients implement this trait.