Trait rusoto_kinesis_video_media::KinesisVideoMedia

pub trait KinesisVideoMedia {
    fn get_media(
        &self, 
        input: GetMediaInput
    ) -> RusotoFuture<GetMediaOutput, GetMediaError>;
}

Trait representing the capabilities of the Kinesis Video Media API. Kinesis Video Media clients implement this trait.

Required methods

fn get_media(
&self,
input: GetMediaInput
) -> RusotoFuture<GetMediaOutput, GetMediaError>

Use this API to retrieve media content from a Kinesis video stream. In the request, you identify the stream name or stream Amazon Resource Name (ARN), and the starting chunk. Kinesis Video Streams then returns a stream of chunks in order by fragment number.

You must first call the GetDataEndpoint API to get an endpoint. Then send the GetMedia requests to this endpoint using the --endpoint-url parameter.

When you put media data (fragments) on a stream, Kinesis Video Streams stores each incoming fragment and related metadata in what is called a "chunk." For more information, see . The GetMedia API returns a stream of these chunks starting from the chunk that you specify in the request.

The following limits apply when using the GetMedia API:

A client can call GetMedia up to five times per second per stream.

Kinesis Video Streams sends media data at a rate of up to 25 megabytes per second (or 200 megabits per second) during a GetMedia session.

Loading content...